+++ native/common/include/jp_booleantype.h
@@ -83,6 +83,10 @@ public:
 	virtual void copyElements(JPJavaFrame &frame,
 			jarray a, jsize start, jsize len,
 			void* memory, int offset) override;
+
+	virtual PyObject *newMultiArray(JPJavaFrame &frame,
+			JPPyBuffer &buffer, int subs, int base, jobject dims) override;
+
 } ;
 
 #endif // _JP_BOOLEAN_TYPE_H_