+++ native/common/include/jp_shorttype.h
@@ -92,6 +92,10 @@ public:
 	virtual void copyElements(JPJavaFrame &frame,
 			jarray a, jsize start, jsize len,
 			void* memory, int offset) override;
+
+	virtual PyObject *newMultiArray(JPJavaFrame &frame,
+			JPPyBuffer &buffer, int subs, int base, jobject dims) override;
+
 } ;
 
 #endif // _JP_SHORT_TYPE_H_