+++ native/common/jp_bytetype.cpp
@@ -265,3 +265,17 @@ void JPByteType::copyElements(JPJavaFrame &frame, jarray a, jsize start, jsize l
 	jbyte* b = (jbyte*) ((char*) memory + offset);
 	frame.GetByteArrayRegion((jbyteArray) a, start, len, b);
 }
+
+static void pack(jbyte* d, jvalue v)
+{
+	*d = v.b;
+}
+
+PyObject *JPByteType::newMultiArray(JPJavaFrame &frame, JPPyBuffer &buffer, int subs, int base, jobject dims)
+{
+	JP_TRACE_IN("JPByteType::newMultiArray");
+	return convertMultiArray<type_t>(
+			frame, this, &pack, "b",
+			buffer, subs, base, dims);
+	JP_TRACE_OUT;
+}