+++ native/common/jp_context.cpp
@@ -233,6 +233,11 @@ void JPContext::startJVM(const string& vmPath, const StringVector& args,
 		m_Context_collectRectangularID = frame.GetMethodID(cls,
 				"collectRectangular",
 				"(Ljava/lang/Object;)[Ljava/lang/Object;");
+
+		m_Context_assembleID = frame.GetMethodID(cls,
+				"assemble",
+				"([ILjava/lang/Object;)Ljava/lang/Object;");
+
 		m_Context_CreateExceptionID = frame.GetMethodID(cls, "createException",
 				"(JJ)Ljava/lang/Exception;");
 		m_Context_GetExcClassID = frame.GetMethodID(cls, "getExcClass",