+++ native/python/include/jp_pythontypes.h
@@ -473,6 +473,8 @@ public:
 	JPPyBuffer(PyObject* obj, int flags);
 	~JPPyBuffer();
 
+	char *getBufferPtr(std::vector<Py_ssize_t>& indices);
+
 	Py_buffer& getView()
 	{
 		return m_View;