+++ doc/CHANGELOG.rst
@@ -4,11 +4,15 @@ Changelog
 This changelog *only* contains changes from the *first* pypi release (0.5.4.3) onwards.
 
 - **Next version - unreleased**
+
   - java.util.Map conforms to Python abc.Mapping API.
 
   - float properly follows Java rules for conversion from double.
     floats outside of range map to inf and -inf.
 
+  - Add support for direct conversion of multi-dimensional primitive arrays
+    with ``JArray.of(array, [dtype=type])``
+    
   - Fix bug that was causing java.lang.Comparable, byte[], and char[] to be unhashable.
 
   - Corrected an issue with creation of Exceptions which lack a